test(tui): organize assistant rendering scenarios
