+++ src/tui/tui-pty-harness.e2e.test.ts
@@ -1,6 +1,10 @@
 // Exercises the fake-backend TUI PTY harness and visible terminal output.
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
 import { sleep } from "../utils/sleep.js";
+import {
+  exerciseDelayedPeerReplies,
+  exerciseLiveReplyRendering,
+} from "./tui-pty-assistant-fixture-test-support.js";
 import { exerciseTuiCommandSurface } from "./tui-pty-command-surfaces-test-support.js";
 import {
   approveWorkspaceSkill,
@@ -12,7 +16,6 @@ import {
   objectFieldEquals,
   readFixtureLog,
   startTuiFixture,
-  waitForSynchronizedFrameRows,
   type FixtureLogEntry,
 } from "./tui-pty-harness-fixture-test-support.js";
 import { registerTuiReconnectTests } from "./tui-pty-reconnect-test-support.js";
@@ -113,53 +116,7 @@ describe.sequential("TUI PTY harness", () => {
 
   it(
     "orders a delayed peer prompt before every retained final reply",
-    async () => {
-      const peerFixture = await startTuiFixture();
-      const markers = [
-        "PTY_PEER_DELAYED_PROMPT",
-        "PTY_PEER_NORMAL_REPLY",
-        "PTY_PEER_MIRROR_REPLY",
-      ] as const;
-      try {
-        await peerFixture.run.waitForOutput("local ready | idle", STARTUP_TIMEOUT_MS);
-        await peerFixture.run.write("/session agent:main:delayed-peer-finals\r", { delay: false });
-        await waitForSynchronizedFrameRows(
-          peerFixture.run,
-          (rows) =>
-            rows.some((row) => row.includes("session agent:main:delayed-peer-finals")) &&
-            rows.some((row) => row.includes("local ready | idle")),
-          STARTUP_TIMEOUT_MS,
-        );
-        await peerFixture.run.write("/gateway-status\r", { delay: false });
-        await peerFixture.waitForLogEntry((entry) => entry.method === "delayedPeerComplete");
-        const rows = await waitForSynchronizedFrameRows(
-          peerFixture.run,
-          (frame) =>
-            markers.every((marker) => frame.some((row) => row.includes(marker))) &&
-            frame.some((row) => row.includes("local ready | idle")),
-          TEST_TIMEOUT_MS,
-        );
-        const entries = await readFixtureLog(peerFixture.logPath);
-        const events = entries.filter((entry) => entry.method === "delayedPeerEvent");
-        const frame = rows.join("\n");
-        console.log(
-          `[behavior-evidence] tui-delayed-peer-finals ${JSON.stringify({
-            rows,
-            events,
-            terminalOutput: peerFixture.run.output(),
-          })}`,
-        );
-        expect(events).toHaveLength(5);
-        expect(entries.some((entry) => entry.method === "sendChat")).toBe(false);
-        for (const marker of markers) {
-          expect(frame.split(marker)).toHaveLength(2);
-        }
-        expect(frame.indexOf(markers[0])).toBeLessThan(frame.indexOf(markers[1]));
-        expect(frame.indexOf(markers[1])).toBeLessThan(frame.indexOf(markers[2]));
-      } finally {
-        await peerFixture.cleanup();
-      }
-    },
+    () => exerciseDelayedPeerReplies(startTuiFixture, STARTUP_TIMEOUT_MS, TEST_TIMEOUT_MS),
     STARTUP_TEST_TIMEOUT_MS,
   );
 
@@ -411,28 +368,7 @@ describe.sequential("TUI PTY harness", () => {
 
   it(
     "renders each live assistant reply once without replaying stale history",
-    async () => {
-      const liveFixture = await startTuiFixture({
-        env: { OPENCLAW_TUI_PTY_COLS: "220", OPENCLAW_TUI_PTY_ROWS: "50" },
-      });
-      try {
-        await liveFixture.run.waitForOutput("local ready", STARTUP_TIMEOUT_MS);
-        await liveFixture.run.write("live reply dedupe proof: first\r", { delay: false });
-        await liveFixture.run.waitForOutput("TUI_LIVE_FIRST");
-        await liveFixture.run.write("live reply dedupe proof: second\r", { delay: false });
-        const rows = await waitForSynchronizedFrameRows(
-          liveFixture.run,
-          (frame) => frame.some((row) => row.includes("TUI_LIVE_SECOND")),
-          STARTUP_TIMEOUT_MS,
-        );
-        const assistantRows = rows.filter(
-          (row) => row.includes("TUI_LIVE_FIRST") || row.includes("TUI_LIVE_SECOND"),
-        );
-        expect(assistantRows).toEqual(["TUI_LIVE_FIRST", "TUI_LIVE_SECOND"]);
-      } finally {
-        await liveFixture.cleanup();
-      }
-    },
+    () => exerciseLiveReplyRendering(startTuiFixture, STARTUP_TIMEOUT_MS),
     STARTUP_TEST_TIMEOUT_MS,
   );
   // prettier-ignore
